fix(debug): use %d not %lld in CAM PPA timing log (nano-printf)
CONFIG_LIBC_NEWLIB_NANO_FORMAT (used by the SeedSigner firmware) has no
64-bit printf conversion, so the CONFIG_CAM_PIPELINE_DEBUG "CAM PPA" line
printed garbage. The microsecond values fit in int; cast and use %d.

+++ src/esp_cam_pipeline.c
@@ -292,9 +292,10 @@ static void frame_cb(uint8_t *camera_buf, uint32_t camera_width,
         if (cam_ppa_dur > cam_ppa_max) cam_ppa_max = cam_ppa_dur;
         cam_ppa_count++;
         if (ppa_t1 - cam_ppa_last_log > 2000000) {  /* every 2s */
-            ESP_LOGI(TAG, "CAM PPA: avg=%lld us  max=%lld us  n=%d",
-                     (long long)(cam_ppa_sum / cam_ppa_count),
-                     (long long)cam_ppa_max, cam_ppa_count);
+            /* %d (not %lld): nano-printf has no 64-bit conversion; us fit in int. */
+            ESP_LOGI(TAG, "CAM PPA: avg=%d us  max=%d us  n=%d",
+                     (int)(cam_ppa_sum / cam_ppa_count),
+                     (int)cam_ppa_max, cam_ppa_count);
             cam_ppa_sum = 0;
             cam_ppa_max = 0;
             cam_ppa_count = 0;
